Updated to become the "official" CollapseDataset module
These are updates so that this can replace the current CollapseDataset
module as the "official" one used by GP.  We have no current source
code for the original, and this one has features needed to support new
workflows around RNA-Seq and modern MSigDB.
Renamed module in the manifest, etc.
Renamed parameters to match older module.
Updated tests with new names.
Added module documentation.  This is still in rough form for now.